+++ source/app/service-providers/config/get-config-template.ts
@@ -124,6 +124,9 @@ export default function getConfigTemplate (): ConfigOptions {
       readabilityAlgorithm: 'dale-chall', // The algorithm to use with readability mode.
       direction: 'ltr', // Can be set to rtl for right-to-left scripts such as Persian
       rtlMoveVisually: true, // Whether the cursor should move visually with arrows in RTL mode
+      lint: {
+        markdown: true // Should Markdown be linted?
+      },
       autoCorrect: {
         active: true, // AutoCorrect is on by default
         style: 'LibreOffice', // Default to LibreOffice style