+++ source/common/modules/markdown-editor/index.ts
@@ -179,6 +179,8 @@ export default class MarkdownEditor extends EventEmitter {
     // eslint-disable-next-line @typescript-eslint/no-this-alias
     const editorInstance = this
 
+    const mdLint = window.config.get('editor.lint.markdown') as boolean
+
     const options: CoreExtensionOptions = {
       initialConfig: JSON.parse(JSON.stringify(this.config)),
       remoteConfig: {
@@ -308,6 +310,9 @@ export default class MarkdownEditor extends EventEmitter {
           }
           return false
         }
+      },
+      lint: {
+        markdown: mdLint
       }
     }